# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spark Lead Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team/members:
    get:
      operationId: getTeamMembers
      tags: [team]
      summary: Get team members
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
  /team/members/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member (role/status)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete team member (full cascade)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /team/invite:
    post:
      operationId: inviteUser
      tags: [team]
      summary: Invite a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserInput"
      responses:
        "201":
          description: Invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /team/users:
    get:
      operationId: getUsersMap
      tags: [team]
      summary: Get user map (id to display name)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserMapEntry"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    TeamMember:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
        whitelistStatus:
          type: string
        joinedAt:
          type: string
      required:
        - id
        - email
        - displayName
        - role
        - whitelistStatus
        - joinedAt
    UpdateTeamMemberInput:
      type: object
      properties:
        role:
          type: string
          enum: [admin, manager, member, lead_owner, deal_handler]
          nullable: true
        status:
          type: string
          enum: [active, disabled]
          nullable: true
    InviteUserInput:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, lead_owner, deal_handler]
      required:
        - email
        - role
    UserMapEntry:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          type: string
      required:
        - id
        - displayName
        - email
        - role
...