# lib/api-zod/src/generated/types/userMapEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark Lead Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface UserMapEntry {
  id: string;
  displayName: string;
  email: string;
  role: string;
}